refactor(settings): move advanced to general
