+++ src/Navigation/Stacks/SettingsStack.tsx
@@ -2,19 +2,18 @@ import React from "react"
 import { createNativeStackNavigator } from "@react-navigation/native-stack"
 import {
     AddContactScreen,
-    AdvancedScreen,
-    ChangeNetworkScreen,
-    ContactsScreen,
     AddCustomNodeScreen,
-    GeneralScreen,
-    PrivacyScreen,
-    SettingsScreen,
-    WalletManagementScreen,
-    ManageCustomNodesScreen,
-    SettingsTransactionsScreen,
-    ManageUrlsScreen,
-    ResetAppScreen,
+    ChangeNetworkScreen,
     ConnectedAppsScreen,
+    ContactsScreen,
+    GeneralScreen,
+    ManageCustomNodesScreen,
+    ManageUrlsScreen,
+    PrivacyScreen,
+    ResetAppScreen,
+    SettingsScreen,
+    SettingsTransactionsScreen,
+    WalletManagementScreen,
 } from "~Screens"
 import { Routes } from "~Navigation/Enums"
 import { AboutScreen } from "~Screens/Flows/App/AboutScreen"
@@ -25,7 +24,6 @@ export type RootStackParamListSettings = {
     [Routes.SETTINGS_ADD_CUSTOM_NODE]: undefined
     [Routes.SETTINGS_MANAGE_CUSTOM_NODES]: undefined
     [Routes.SETTINGS_PRIVACY]: undefined
-    [Routes.SETTINGS_ADVANCED]: undefined
     [Routes.SETTINGS_ABOUT]: undefined
     [Routes.SETTINGS_GENERAL]: undefined
     [Routes.SETTINGS_CONTACTS]: undefined
@@ -72,11 +70,6 @@ export const SettingsStack = () => {
                     options={{ headerShown: false }}
                 />
 
-                <Settings.Screen
-                    name={Routes.SETTINGS_ADVANCED}
-                    component={AdvancedScreen}
-                    options={{ headerShown: false }}
-                />
                 <Settings.Screen
                     name={Routes.SETTINGS_ABOUT}
                     component={AboutScreen}
