--- src/Screens/Flows/App/AdvancedScreen/index.ts
@@ -1,1 +1,0 @@
-export * from "./AdvancedScreen"
