+++ src/Screens/Flows/App/GeneralScreen/GeneralScreen.tsx
@@ -14,16 +14,19 @@ import {
     SelectLanguageBottomSheet,
 } from "./Components"
 import { useBottomSheetModal } from "~Hooks"
-import { isSmallScreen, LANGUAGE } from "~Constants"
+import { LANGUAGE } from "~Constants"
 import {
     selectAreDevFeaturesEnabled,
     selectHideTokensWithNoBalance,
     selectLangauge,
+    selectSentryTrackingEnabled,
     setHideTokensWithNoBalance,
     setLanguage,
+    setSentryTrackingEnabled,
     useAppDispatch,
     useAppSelector,
 } from "~Storage/Redux"
+import { Reset } from "~Screens/Flows/App/GeneralScreen/Components/Reset"
 
 export const GeneralScreen = () => {
     const { LL } = useI18nContext()
@@ -38,6 +41,15 @@ export const GeneralScreen = () => {
 
     const selectedLanguage = useAppSelector(selectLangauge)
     const devFeaturesEnabled = useAppSelector(selectAreDevFeaturesEnabled)
+
+    const sentryTrackingEnabled = useAppSelector(selectSentryTrackingEnabled)
+
+    const toggleSentryTrackingSwitch = useCallback(
+        (newValue: boolean) => {
+            dispatch(setSentryTrackingEnabled(newValue))
+        },
+        [dispatch],
+    )
 
     const hideTokensWithNoBalance = useAppSelector(
         selectHideTokensWithNoBalance,
@@ -62,7 +74,7 @@ export const GeneralScreen = () => {
     return (
         <Layout
             safeAreaTestID="General_Screen"
-            isScrollEnabled={isSmallScreen}
+            isScrollEnabled={devFeaturesEnabled}
             body={
                 <BaseView pt={16}>
                     <BaseText typographyFont="title">
@@ -94,6 +106,29 @@ export const GeneralScreen = () => {
                     <BaseSpacer height={20} />
 
                     <ChangeTheme />
+
+                    <BaseSpacer height={24} />
+                    <BaseText typographyFont="bodyMedium" my={8}>
+                        {LL.BD_RESET()}
+                    </BaseText>
+                    <BaseText typographyFont="caption">
+                        {LL.BD_RESET_DISCLAIMER()}
+                    </BaseText>
+
+                    <BaseSpacer height={16} />
+                    <Reset />
+
+                    {devFeaturesEnabled && (
+                        <>
+                            <BaseSpacer height={24} />
+                            <EnableFeature
+                                title={LL.BD_HELP_IMPROVE()}
+                                subtitle={LL.BD_HELP_IMPROVE_DISCLAIMER()}
+                                onValueChange={toggleSentryTrackingSwitch}
+                                value={sentryTrackingEnabled}
+                            />
+                        </>
+                    )}
 
                     <BaseSpacer height={20} />
