+++ src/Screens/Flows/App/index.ts
@@ -6,7 +6,6 @@ export * from "./NetworkScreen"
 export * from "./AddCustomNodeScreen"
 export * from "./ManageCustomNodesScreen"
 export * from "./PrivacyScreen"
-export * from "./AdvancedScreen"
 export * from "./GeneralScreen"
 export * from "./BuyScreen"
 export * from "./SendScreen"
